User selection without UI for it
I'm too lazy to render the UI for the User detail view, but this is one way of hooking it all up. It's a bit roundabout with that ViewModel MutableStateFlow for the selected User: things could be simpler with eliminating that middleman, especially since this is a one click and then we're done deal: so the Activity would just pass in its own click listener that it would use to navigate. Is the way I have it more "official"? Probably so.

I'm parceling the User because it's easier and fun. I suppose it's not completely maintainable because eventually our User object could get too big to parcel nicely. We could also go fetch the User from our database. To do that, we'd parcel the id of the User, add to our Room db the ability to fetch a single User by id -- a simple suspend fun will do it (no need for a Flow here), add a method to our repository to call that db method, and use a LifecycleScope to call it/scope it from our Activity (or add a ViewModel method to do that and have our Activity call that ViewModel proxy?)

Fetching user detail via the database now.

Just a sketch still, and this could be made even more correct (as explained in my code comments) by splitting up the scoping of the fetching of the User from the observation of when that fetch is done.